+++ changelog.md
@@ -12,6 +12,9 @@
 - Add debug hud to show enchantment tags for currently held item
   - Access via `F3+F6`
   - Named `autoswitch:enchantment_tags`
+- Add debug hud to show entity tags of the currently targeted entity
+    - Access via `F3+F6`
+    - Named `autoswitch:entity_tags`
 
 # 13.1.0
 - Implement enchantment level specification for item selectors